requirements.txt
streamlit
rembg[gpu]==2.0.57
streamlit-cropper
Pillow
onnxruntime
